Don't use -fstack-protector-strong when compiling for Windows
Initial motivation of adding the flag was better consistency with
compile flags used for regular Arch Linux packages.

However, adding -fstack-protector-strong seems to cause segmentation
faults even before main() is called.

The issue is not reproducible with a simple hello world app and also
not reproducible when launching the app under Linux with WINE. But
a full-blown Qt 5 app crashes in the way described.