Add separate packages for static Qt 5 modules (mingw-w64)
The static Qt modules in the base repository are already packaged as a
separate package (mingw-w64-qt5-base-static). This change applies the same
separation to Qt modules found in further repositories. That means it is no
longer required to build a static version of Qt if only shared libraries
are required.

Note that there's no static variant for mingw-w64-qt5-activeqt because this
module ends up being built as a static library anyways. There is also no
static version of mingw-w64-qt5-translations because it does not contain
any libraries.