Downgrade libx11 and add shared version
We must not link statically against it because it is also pulled by the
shared GLX library and would therefore end up multiple times in the apps
address space causing crashes. It should be stable enough anyways to not
having to bundle it. Since 1.6.x is still the newest version on older
distributions we must downgrade the version, though.