Drop gcc-d-static-compat in accordance with regular gcc package
It wouldn't be possible to keep it nevertheless because then
we needed to add gcc-d to make dependencies which is not possible
because the regular package dropped it. It is also not possible
to build with the previous version of static-compat-gcc because
it would fail to find gmp and some other libs (using them from
the host system would be technically possible but one needed to
workaround linker errors with the C runtime).

For some reason also libobjc is missing but I suppose that's not
a bit loss.