Use temporary directory as HOME for syncthingtray tests
Some part of the syncthingtray or syncthing tests creates a Sync
directory under $HOME if it doesn't exist.  This can be annoying for
users building or testing the package when they don't use the default
$HOME/Sync directory as it leaves behind test artifacts in their home.

Use a temporary directory in the source tree when testing instead.
Don't export the variable in case the overridden $HOME leaks into other
parts of the build process.