Ensure mingw-w64-clang-aarch64-winpthreads is available
Many packages probably need it now that the headers are not present anymore
in mingw-w64-clang-aarch64-headers. This change simply makes
mingw-w64-clang-aarch64-compiler-rt depend on it so it'll be pulled into
the build root of all mingw-w64-clang-aarch64 packages but not into the
build root of headers and crt.