Add hack so CMake does not omit static libraries
When linking against libarchive statically and therefore also against its
transitive dependency bz2 CMake omits bz2 for freetype2 for some reason.
This is wrong and leads to linker errors. Not sure why that is the case.
Maybe it is even a CMake bug. Using a rescan link group fortunately helps
to circumvent this issue.

Note that setting `IMPORTED_LINK_INTERFACE_MULTIPLICITY` or
`LINK_INTERFACE_MULTIPLICITY` doesn't help (which is expected as those
limits are set to allow two occurrences of the same library by default
anyway but here CMake even limits the number to just one occurrence which
must be for some different reason).