Avoid using `-Wl,-z,pack-relative-relocs` in static-compat env
This linker flag enables the DT_RELR relative relocation format which is
only supported by glibc 2.36 and thus using it breaks compatibility with
older distributions.

This flag should stay of course in the general build env keeping it in-line
with the flags used by Arch Linux and only be omitted when building static-
compat packaging.

See https://sourceware.org/bugzilla/show_bug.cgi?id=27924 for the issue
about this feature on the glibc bugtracker.