Revert Qt 6 commit that causes problems on Android Samsung device
With this commit the Android activity can get in a state where it no
longer reacts to input events and instead just logs "Touch event received
for non-existing window 1" when one occurs.

I could only reproduce this on a real Samsung device so far but not in the
emulator. On the Samsung device the problem occurs after the activity is
destroyed. When it is then reopened again a flickering occurs and then the
UI ends up in this bad state.